refactor: move loading logic back into llmb, simplify
