+++ crates/llm-base/src/model/mod.rs
@@ -4,7 +4,7 @@ use std::{fmt::Debug, path::PathBuf, sync::Arc};
 
 use ggml::{
     accelerator::Backend,
-    format::gguf::{Metadata, MetadataError},
+    format::gguf::{Gguf, MetadataError},
     sys::llama::llama_ftype,
 };
 use regex::Regex;
@@ -13,26 +13,47 @@ use thiserror::Error;
 use crate::{
     loader::{ModelTensorLoader, TensorLoadError},
     tokenizer::TokenId,
-    FileType, InferenceSession, InferenceSessionConfig, Tokenizer,
+    InferenceSession, InferenceSessionConfig, Tokenizer,
 };
 
 /// Common functions for model evaluation
 pub mod common;
 
+/// All of the arguments required to load a model.
+pub struct ModelLoadArgs<'a> {
+    /// The GGUF metadata for the model.
+    pub gguf: &'a Gguf,
+    /// Model metadata.
+    pub data: ModelData,
+    /// The tensor loader to use for the model.
+    pub tensor_loader: ModelTensorLoader<'a>,
+}
+
+/// Model data that is required for all models.
+pub struct ModelData {
+    /// Any parameters that control the behaviour of the model.
+    pub params: ModelParameters,
+    /// The tokenizer to use for the model.
+    pub tokenizer: Tokenizer,
+}
+
+/// An error encountered while loading a concrete model.
+#[derive(Error, Debug)]
+pub enum ModelLoadError {
+    /// An error occurred while loading the model's tensors.
+    #[error("{0}")]
+    TensorLoadError(#[from] TensorLoadError),
+    /// An error occurred while reading the model's hyperparameters.
+    #[error("{0}")]
+    HyperparametersReadError(#[from] HyperparametersReadError),
+}
+
 /// Interfaces for creating and interacting with a large language model with a known type
 /// of [hyperparameters](https://en.wikipedia.org/wiki/Hyperparameter_(machine_learning)).
-pub trait KnownModel: Send + Sync {
-    /// Hyperparameters for the model.
-    type Hyperparameters: Hyperparameters;
-
+pub trait Model: Send + Sync {
     /// Creates a new model from the provided [ModelParameters] hyperparameters.
     /// This function is called by the [load](crate::loader::load) function.
-    fn new(
-        hyperparameters: Self::Hyperparameters,
-        params: ModelParameters,
-        tokenizer: Tokenizer,
-        tensor_loader: ModelTensorLoader,
-    ) -> Result<Self, TensorLoadError>
+    fn new(args: ModelLoadArgs) -> Result<Self, ModelLoadError>
     where
         Self: Sized;
 
@@ -50,15 +71,19 @@ pub trait KnownModel: Send + Sync {
         output_request: &mut OutputRequest,
     );
 
-    /// Get the hyperparameters for this model.
-    fn hyperparameters(&self) -> &Self::Hyperparameters;
+    /// Get the data for this model.
+    fn data(&self) -> &ModelData;
 
     /// Get the tokenizer for this model.
-    fn tokenizer(&self) -> &Tokenizer;
+    fn tokenizer(&self) -> &Tokenizer {
+        &self.data().tokenizer
+    }
 
     /// Get the context size (configured with [ModelParameters::context_size]) used by
     /// this model.
-    fn context_size(&self) -> usize;
+    fn context_size(&self) -> usize {
+        self.data().params.context_size
+    }
 
     /// Get the beginning of text/beginning of string token ID, if available. This value is defined by model implementers.
     fn bot_token_id(&self) -> Option<TokenId>;
@@ -67,10 +92,10 @@ pub trait KnownModel: Send + Sync {
     fn eot_token_id(&self) -> TokenId;
 
     /// Get the list of regexes to use to determine if a tensor in this model should be quantized.
-    fn quantize_tensors() -> Vec<Regex>;
+    fn quantize_tensors(&self) -> Vec<Regex>;
 
     /// Get the list of regexes to use to determine if a tensor in this model should not be quantized.
-    fn skip_quantize_tensors() -> Vec<Regex>;
+    fn skip_quantize_tensors(&self) -> Vec<Regex>;
 
     /// Returns whether the model supports deleting tokens.
     fn supports_rewind(&self) -> bool {
@@ -79,91 +104,8 @@ pub trait KnownModel: Send + Sync {
     }
 }
 
-/// A type-erased model to allow for interacting with a model without knowing
-/// its hyperparameters.
-pub trait Model: Send + Sync {
-    /// Starts a new `InferenceSession` for this model.
-    fn start_session(&self, config: InferenceSessionConfig) -> InferenceSession;
-
-    /// This function is called by the provided [InferenceSession]; it will use this model
-    /// to generate output by evaluating the `input_tokens`.
-    /// The [OutputRequest] is used to specify additional data to fetch from the
-    /// model.
-    fn evaluate(
-        &self,
-        session: &mut InferenceSession,
-        input_tokens: &[TokenId],
-        output_request: &mut OutputRequest,
-    );
-
-    /// Get the tokenizer for this model.
-    fn tokenizer(&self) -> &Tokenizer;
-
-    /// Get the context size (configured with [ModelParameters::context_size]) used by
-    /// this model.
-    fn context_size(&self) -> usize;
-
-    /// Get the beginning of text/beginning of string token ID, if available. This value is defined by model implementers.
-    fn bot_token_id(&self) -> Option<TokenId>;
-
-    /// Get the end of text/end of string token ID. This value is defined by model implementers.
-    fn eot_token_id(&self) -> TokenId;
-
-    /// Returns whether the model supports deleting tokens.
-    fn supports_rewind(&self) -> bool;
-}
-impl<H: Hyperparameters, M: KnownModel<Hyperparameters = H>> Model for M {
-    fn start_session(&self, config: InferenceSessionConfig) -> InferenceSession {
-        KnownModel::start_session(self, config)
-    }
-
-    fn evaluate(
-        &self,
-        session: &mut InferenceSession,
-        input_tokens: &[TokenId],
-        output_request: &mut OutputRequest,
-    ) {
-        KnownModel::evaluate(self, session, input_tokens, output_request)
-    }
-
-    fn tokenizer(&self) -> &Tokenizer {
-        KnownModel::tokenizer(self)
-    }
-
-    fn context_size(&self) -> usize {
-        KnownModel::context_size(self)
-    }
-
-    fn bot_token_id(&self) -> Option<TokenId> {
-        KnownModel::bot_token_id(self)
-    }
-
-    fn eot_token_id(&self) -> TokenId {
-        KnownModel::eot_token_id(self)
-    }
-
-    fn supports_rewind(&self) -> bool {
-        KnownModel::supports_rewind(self)
-    }
-}
-
-/// Implemented by model hyperparameters for interacting with hyperparameters
-/// without knowing what they are, as well as writing/reading them as required.
-pub trait Hyperparameters: Sized + Default + Debug + PartialEq + Eq {
-    /// Read the parameters from GGUF metadata.
-    fn read_gguf(metadata: &Metadata) -> Result<Self, HyperparametersReadError>;
-
-    /// Write the parameters to GGUF metadata.
-    fn write_gguf(&self, metadata: &mut Metadata) -> Result<(), HyperparametersWriteError>;
-
-    /// Get the filetype of the model.
-    fn file_type(&self) -> Option<FileType>;
-
-    /// Get mutable access to filetype of the model.
-    fn file_type_mut(&mut self) -> Option<&mut FileType>;
-}
 #[derive(Error, Debug)]
-/// Reported from functions that write
+/// Reported from functions that read hyperparameters
 pub enum HyperparametersReadError {
     #[error("{0}")]
     /// A metadata error.
@@ -174,16 +116,6 @@ pub enum HyperparametersReadError {
         /// The file type (ignoring the quantization version) that was encountered.
         file_type: llama_ftype,
     },
-}
-#[derive(Error, Debug)]
-/// Reported from functions that write
-pub enum HyperparametersWriteError {
-    #[error("non-specific I/O error")]
-    /// A non-specific IO error.
-    Io(#[from] std::io::Error),
-    #[error("invalid integer conversion")]
-    /// One of the integers encountered could not be converted to a more appropriate type.
-    InvalidIntegerConversion(#[from] std::num::TryFromIntError),
 }
 
 /// Parameters for model-wide behaviour.
