+++ crates/models/gptj/src/lib.rs
@@ -7,7 +7,7 @@
 // use llm_base::{
 //     ggml,
 //     model::{common, HyperparametersWriteError},
-//     util, FileType, GraphOutputs, InferenceSession, InferenceSessionConfig, KnownModel, LoadError,
+//     util, FileType, GraphOutputs, InferenceSession, InferenceSessionConfig, Model, LoadError,
 //     ModelContext, ModelParameters, OutputRequest, Regex, TensorLoader, TokenId, Tokenizer,
 // };
 
@@ -41,7 +41,7 @@
 // unsafe impl Send for GptJ {}
 // unsafe impl Sync for GptJ {}
 
-// impl KnownModel for GptJ {
+// impl Model for GptJ {
 //     type Hyperparameters = Hyperparameters;
 
 //     fn new<E: Error>(
